added the support of n-dimension array from branch 'dataspaces-multi-dims' revision 1522.

+++ include/sfc.h
@@ -35,9 +35,9 @@
 #define _hilbert_h_
 
 /* define the bitmask_t type as an integer of sufficient size */
-typedef uint64_t bitmask_t;
+typedef unsigned long long bitmask_t;
 /* define the halfmask_t type as an integer of 1/2 the size of bitmask_t */
-typedef uint32_t halfmask_t;
+typedef unsigned long halfmask_t;
 
 #define adjust_rotation(rotation,nDims,bits)                            \
 do {                                                                    \
